Moved parse! function to the BaumeisterFile module
