+++ apps/baumeister/test/baumeister_test.exs
@@ -1,6 +1,7 @@
 defmodule BaumeisterTest do
   use ExUnit.Case
   doctest Baumeister
+  doctest Baumeister.BaumeisterFile
 
   test "the truth" do
     assert 1 + 1 == 2
